added voting to squads that updates on all screens
So now you can select your avatar and it will be shown when voting for a squad. Your cursor is a number between 0 and 3 based on a a random number assigned to you when creating  starting the game (later cursors can be changed to real cursors not only showing the numbers).

It was a hassle to make it work but I'm happy with the results :)